add force sync button and autorestart tracing over reset
