+++ app/src/main/res/values-de/strings.xml
@@ -396,6 +396,9 @@ Corona-Infoline
     <!-- Fuzzy -->
     <string name="debug_settings_title">"Debug Einstellungen"</string>
 
+    <string name="debug_force_sync_title">"Force sync"</string>
+    <string name="debug_force_sync_subtitle">"set sync period at 10s"</string>
+
     <!-- Titel für die Debug-Einstellungen zur Status-Überschreibung -->
     <string name="debug_state_setting_title">"App-Zustand überschreiben"</string>
